simple-rag/requirements.txt
langchain-chroma
langchain-openai
python-dotenv
pydantic-settings
langchain-core
langchain
PyMuPDF
openai